Prevent (signed) integer overflows in memorystats code
Patch from @MPC7500 in #745 - thanks!

# Conflicts:
#	neo/framework/Console.cpp

+++ neo/idlib/Heap.h
@@ -46,7 +46,7 @@ typedef struct {
 	int		num;
 	int		minSize;
 	int		maxSize;
-	int		totalSize;
+	long long totalSize; // DG: prevent overflows by using a 64bit int for totalSize
 } memoryStats_t;
